added a readme creator, built a readme class that takes all the necessay info as its constructor and contains a method that writes that info to a markdown file template.
