+++ package.json
@@ -1,0 +1,27 @@
+{
+  "name": "playingwithjavascriptclasses",
+  "version": "1.0.0",
+  "description": "a collection of small projects built using classes in JavaScript",
+  "main": "house.js",
+  "directories": {
+    "test": "tests"
+  },
+  "dependencies": {
+    "inquirer": "^9.1.0",
+    "jest": "^28.1.3",
+    "typescript": "^4.7.4"
+  },
+  "scripts": {
+    "test": "jest"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/jdlevish/playingWithJavascriptClasses.git"
+  },
+  "author": "Joshua Levish",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/jdlevish/playingWithJavascriptClasses/issues"
+  },
+  "homepage": "https://github.com/jdlevish/playingWithJavascriptClasses#readme"
+}
